Add files via upload
This project gives details about the top cities as well as the ranking of the retailers by revenue. Charleston city brought in the highest revenue and the retailer with the highest revenue line was West Gear. 
The shopping methods (in-store, online and outlet) were also analysed; the in-store shopping method generated the highest sales.
The highest-grossing products were also accounted for in the project.